+++ src/base64.ts
@@ -1,6 +1,5 @@
-import { log } from "@dwtechs/winstan";
 import { isString, isBase64 } from "@dwtechs/checkard";
-import { LOGS_PREFIX } from "./constants";
+import { InvalidBase64FormatError, InvalidStringForEncodingError } from "./errors";
 
 /**
  * Decodes a base64 encoded string.
@@ -8,12 +7,17 @@ import { LOGS_PREFIX } from "./constants";
  * @param {string} str - The base64 encoded string to decode.
  * @param {boolean} urlSafe - A boolean indicating if the input string is URL safe. Defaults to true.
  * @returns {string} The decoded string in UTF-8 format.
- * @throws {Error} If `str` is not a valid base64 string.
+ * @throws {InvalidBase64FormatError} If `str` is not a valid base64 string.
  */
 function b64Decode(str: string, urlSafe = true): string {
-  log.debug(`${LOGS_PREFIX}Decoding base64 string (urlSafe=${urlSafe})`);
   
-  isBase64(str, urlSafe, true);
+  try {
+    isBase64(str, urlSafe, true);
+  } catch (err) {
+    const e = new InvalidBase64FormatError(urlSafe);
+    e.cause = err;
+    throw e;
+  }
 
   if (urlSafe)
     str = str.replace(/-/g, "+").replace(/_/g, "/");
@@ -29,11 +33,18 @@ function b64Decode(str: string, urlSafe = true): string {
  * @param {string} str - The string to be encoded.
  * @param {boolean} urlSafe - Optional boolean to determine if the output should be URL safe. Defaults to true.
  * @returns {string} The Base64 encoded string. If `urlSafe` is true, the output will be modified to be URL safe.
- * @throws {Error} If `str` is not a non-empty string.
+ * @throws {InvalidStringForEncodingError} If `str` is not a non-empty string.
  */
 function b64Encode(str: string, urlSafe = true): string {
-  log.debug(`${LOGS_PREFIX}Encoding string (urlSafe=${urlSafe})`);
-  isString(str, "!0", null, true);
+
+  try {
+    isString(str, "!0", null, true);
+  } catch (err) {
+    const e = new InvalidStringForEncodingError();
+    e.cause = err;
+    throw e;
+  }
+  
   let b64 = Buffer.from(str).toString("base64");
   if (urlSafe)
     b64 = b64.replace(/\+/g, "-")